Remove randomness from vessel generation
