+++ README.md
@@ -5,3 +5,7 @@ This prototype demonstrates a basic browser-based simulator for guiding a stiff 
 ## Usage
 
 Open `index.html` in a modern browser. Use `W`/`S` or the up/down arrow keys to advance or retract the guidewire through the introducer sheath positioned in the left branch. The push distance now allows the wire to be fully inserted if desired.
+
+## Vessel Geometry
+
+The vessel is generated deterministically. Branch length and angle offset use fixed defaults (140 units and 0 radians) and only change when explicitly provided to `generateVessel`.
